refactor(notify): Phase 3.3 — extract notify_scan + alert logic to backend/notify/
Pure move: _post_json, _post_text, _tg_escape, _alert_host_label,
dispatch_alert, _dispatch_to_channels, _alert_name, notify_scan moved to
backend/notify/__init__.py with lazy _app imports. Intra-module calls use
_app.* so test patches on app.* continue to work. app.py re-exports all
moved names. Snapshot suite confirms API responses unchanged.
