+++ app.py
@@ -4203,19 +4203,11 @@ _NTFY_T = {"info": "information_source", "warning": "warning", "critical": "rota
 # also mandated by Discord's API rules, so every outbound POST carries one.
 NOTIFY_USER_AGENT = f"homelab-monitor/{VERSION} (+https://github.com/SikamikanikoBG/homelab-monitor)"
 
-def _post_json(url, payload, timeout=5):
-    req = urllib.request.Request(url, data=json.dumps(payload).encode("utf-8"),
-                                 headers={"Content-Type": "application/json",
-                                          "User-Agent": NOTIFY_USER_AGENT})
-    with urllib.request.urlopen(req, timeout=timeout) as r:
-        return r.status, r.read()
+# Phase 3.3: moved to backend/notify/ — re-exported for backward compat
+from backend.notify import _post_json
 
-def _post_text(url, text, headers=None, timeout=5):
-    hdr = dict(headers or {"Content-Type": "text/plain"})
-    hdr.setdefault("User-Agent", NOTIFY_USER_AGENT)
-    req = urllib.request.Request(url, data=text.encode("utf-8"), headers=hdr)
-    with urllib.request.urlopen(req, timeout=timeout) as r:
-        return r.status, r.read()
+# Phase 3.3: moved to backend/notify/ — re-exported for backward compat
+from backend.notify import _post_text
 
 def _post_multipart(url, payload_json, filename, file_bytes,
                     file_ctype="text/html; charset=utf-8", timeout=10):
@@ -4271,10 +4263,8 @@ def send_ntfy(server, topic, level, title, detail):
               "Tags":     _NTFY_T.get(level, "information_source")}
     return _post_text(url, detail, hdr)
 
-def _tg_escape(text):
-    """Escape Telegram legacy-Markdown metacharacters in user-supplied text."""
-    return (text or "").replace("\\", "\\\\").replace("_", "\\_").replace("*", "\\*") \
-                       .replace("`", "\\`").replace("[", "\\[")
+# Phase 3.3: moved to backend/notify/ — re-exported for backward compat
+from backend.notify import _tg_escape
 
 def _post_to_telegram(token, chat_id, level, title, body):
     url = f"https://api.telegram.org/bot{token}/sendMessage"
@@ -4318,92 +4308,14 @@ def send_webhook(url, level, title, detail, host):
     payload = {"level": level, "title": title, "detail": detail, "host": host}
     return _post_json(url, payload)
 
-def _alert_host_label():
-    """Machine name to stamp on every alert so a notification says *where* the
-    problem is. Alerts are raised from the hub's own docker/systemd/disk/GPU
-    snapshots, so this is the hub host: prefer the probe-reported hostname (the
-    same name the dashboard's host tab shows), fall back to the OS hostname, and
-    finally to "" so a label-less environment degrades to the old behaviour."""
-    try:
-        name = ((LATEST or {}).get("host") or {}).get("hostname")
-        if name:
-            return str(name).strip()
-    except Exception:
-        pass
-    try:
-        return socket.gethostname()
-    except Exception:
-        return ""
+# Phase 3.3: moved to backend/notify/ — re-exported for backward compat
+from backend.notify import _alert_host_label
 
-def dispatch_alert(s, level, title, detail, host=None):
-    """Send to whichever channels are configured. Returns list of (channel, ok, err).
+# Phase 3.3: moved to backend/notify/ — re-exported for backward compat
+from backend.notify import dispatch_alert
 
-    `title` is prefixed with the machine name (`[host] …`) so every channel —
-    Discord, ntfy, Telegram, email, Slack and generic webhook alike — names
-    which machine the alert is about.
-    Pass host="" to opt out (e.g. a generic message that isn't host-specific)."""
-    if host is None:
-        host = _alert_host_label()
-    if host:
-        title = f"[{host}] {title}"
-    out = []
-    if s.get("discord_webhook_url"):
-        try: send_discord(s["discord_webhook_url"], level, title, detail); out.append(("discord", True, None))
-        except Exception as e: out.append(("discord", False, str(e)))
-    if s.get("ntfy_topic"):
-        try: send_ntfy(s.get("ntfy_server") or "https://ntfy.sh",
-                       s["ntfy_topic"], level, title, detail); out.append(("ntfy", True, None))
-        except Exception as e: out.append(("ntfy", False, str(e)))
-    if s.get("telegram_token") and s.get("telegram_chat_id"):
-        try: _post_to_telegram(s["telegram_token"], s["telegram_chat_id"],
-                               level, title, detail); out.append(("telegram", True, None))
-        except Exception as e: out.append(("telegram", False, str(e)))
-    # Email via SMTP
-    if s.get("email_host") and s.get("email_from") and s.get("email_to"):
-        try:
-            _send_email(s["email_host"], s.get("email_port", "587"),
-                        s.get("email_use_tls", "1") == "1",
-                        s.get("email_username", ""), s.get("email_password", ""),
-                        s["email_from"], s["email_to"],
-                        level, title, detail)
-            out.append(("email", True, None))
-        except Exception as e:
-            out.append(("email", False, str(e)))
-    # Slack incoming webhook
-    if s.get("slack_webhook_url"):
-        try: send_slack(s["slack_webhook_url"], level, title, detail); out.append(("slack", True, None))
-        except Exception as e: out.append(("slack", False, str(e)))
-    # Generic webhook
-    if s.get("webhook_url"):
-        try: send_webhook(s["webhook_url"], level, title, detail, host or ""); out.append(("webhook", True, None))
-        except Exception as e: out.append(("webhook", False, str(e)))
-    return out
-
-def _dispatch_to_channels(s, level, title, detail, channels):
-    """Dispatch to a specific set of channels only."""
-    if "discord" in channels and s.get("discord_webhook_url"):
-        try: send_discord(s["discord_webhook_url"], level, title, detail)
-        except Exception as e: print("notifier discord error:", e, flush=True)
-    if "ntfy" in channels and s.get("ntfy_topic"):
-        try: send_ntfy(s.get("ntfy_server") or "https://ntfy.sh",
-                       s["ntfy_topic"], level, title, detail)
-        except Exception as e: print("notifier ntfy error:", e, flush=True)
-    if "telegram" in channels and s.get("telegram_token") and s.get("telegram_chat_id"):
-        try: _post_to_telegram(s["telegram_token"], s["telegram_chat_id"],
-                               level, title, detail)
-        except Exception as e: print("notifier telegram error:", e, flush=True)
-    if "email" in channels and s.get("email_host") and s.get("email_from") and s.get("email_to"):
-        try: _send_email(s["email_host"], s.get("email_port", "587"),
-                         s.get("email_use_tls", "1") == "1",
-                         s.get("email_username", ""), s.get("email_password", ""),
-                         s["email_from"], s["email_to"], level, title, detail)
-        except Exception as e: print("notifier email error:", e, flush=True)
-    if "slack" in channels and s.get("slack_webhook_url"):
-        try: send_slack(s["slack_webhook_url"], level, title, detail)
-        except Exception as e: print("notifier slack error:", e, flush=True)
-    if "webhook" in channels and s.get("webhook_url"):
-        try: send_webhook(s["webhook_url"], level, title, detail, _alert_host_label() or "")
-        except Exception as e: print("notifier webhook error:", e, flush=True)
+# Phase 3.3: moved to backend/notify/ — re-exported for backward compat
+from backend.notify import _dispatch_to_channels
 
 def _match_kind(key):
     prefix = key.split(":")[0] if ":" in key else key
@@ -4415,9 +4327,8 @@ def _match_kind(key):
         return "host"
     return "host"
 
-def _alert_name(key):
-    parts = key.split(":")
-    return ":".join(parts[1:]) if len(parts) > 1 else key
+# Phase 3.3: moved to backend/notify/ — re-exported for backward compat
+from backend.notify import _alert_name
 
 def _apply_rules(key, level, rules):
     """Return set of channels if rules match, or None for default behavior."""
@@ -4527,109 +4438,8 @@ def delete_maintenance_window(wid):
         DB.execute("DELETE FROM maintenance_windows WHERE id=?", (wid,))
         DB.commit()
 
-def notify_scan():
-    s = get_settings()
-    if s.get("alerts_enabled") != "1":
-        return
-    if not (s.get("discord_webhook_url") or s.get("ntfy_topic")
-            or (s.get("telegram_token") and s.get("telegram_chat_id"))
-            or (s.get("email_host") and s.get("email_from") and s.get("email_to"))
-            or s.get("slack_webhook_url")
-            or s.get("webhook_url")):
-        return
-    rules = get_notification_rules()
-
-    # ── Docker containers: edge-trigger on crit/warn, clear on ok ─────────────
-    docker = HEALTH.get("docker") or {}
-    if docker.get("available"):
-        for ct in docker.get("containers", []):
-            name = ct.get("name", "?")
-            key  = f"container:{name}"
-            st   = ct.get("status")
-            if st == "crit":
-                _emit(s, key, "critical", f"🔴 Container {name} {ct.get('label','')}".strip(),
-                      f"{name}: {ct.get('status_text','')}", rules=rules)
-            elif st == "warn":
-                _emit(s, key, "warning", f"🟠 Container {name} {ct.get('label','')}".strip(),
-                      f"{name}: {ct.get('status_text','')}", rules=rules)
-            elif st == "ok":
-                _clear(key)
-
-    # ── systemd units: edge-trigger on failed ─────────────────────────────────
-    systemd = HEALTH.get("systemd") or {}
-    if systemd.get("available"):
-        for svc in systemd.get("services", []):
-            name = svc.get("name", "?")
-            key  = f"systemd:{name}"
-            if svc.get("status") == "crit":
-                _emit(s, key, "critical", f"🔴 systemd unit failed: {name}",
-                      f"{name} — {svc.get('desc','')} (active={svc.get('active')}, sub={svc.get('sub')})",
-                      rules=rules)
-            elif svc.get("status") == "ok":
-                _clear(key)
-
-    # ── GPU VRAM pressure ────────────────────────────────────────────────────
-    mem_total = LATEST.get("mem_total") or 0
-    mem_used  = LATEST.get("mem_used")  or 0
-    if mem_total:
-        free = mem_total - mem_used
-        key  = "gpu:vram_pressure"
-        if free < PRESSURE_MB:
-            _emit(s, key, "warning", "🟠 GPU VRAM pressure",
-                  f"Only {round(free)} MB free of {round(mem_total)} MB "
-                  f"({round(100*mem_used/mem_total)}% used).", rules=rules)
-        else:
-            _clear(key)
-
-    # ── Disks crossing the configured threshold ───────────────────────────────
-    try: disk_thr = int(s.get("disk_alert_pct") or 90)
-    except ValueError: disk_thr = 90
-    host = LATEST.get("host") or {}
-    seen_disks = set()
-    for dk in (host.get("disks") or []):
-        mp   = dk.get("mount", "?")
-        seen_disks.add(mp)
-        key  = f"disk:{mp}"
-        pct  = dk.get("pct", 0)
-        if pct >= disk_thr:
-            level = "critical" if pct >= 95 else "warning"
-            _emit(s, key, level, f"{'🔴' if level=='critical' else '🟠'} Disk {mp} at {pct}%",
-                  f"{mp}: {dk.get('used',0)} GB / {dk.get('total',0)} GB used ({pct}%).",
-                  rules=rules)
-        else:
-            _clear(key)
-
-    # ── GPU OOM events from the DB (each event_ts notified at most once) ─────
-    try:
-        cutoff = int(time.time()) - 3600
-        with LOCK:
-            rows = DB.execute("SELECT ts, service, detail FROM events "
-                              "WHERE kind='oom' AND ts>=? ORDER BY ts", (cutoff,)).fetchall()
-        for ets, svc, detail in rows:
-            key = f"oom:{svc}:{ets}"
-            with _NOTIFIER_LOCK:
-                already = key in _NOTIFIED
-                if not already:
-                    _NOTIFIED[key] = 1
-            if already:
-                continue
-            if LEVELS["critical"] < LEVELS.get(s.get("alert_min_level", "warning"), 1):
-                continue
-            channels = _apply_rules(key, "critical", rules)
-            if channels is not None:
-                _dispatch_to_channels(s, "critical", f"🔴 GPU OOM in {svc}", (detail or "")[:1500], channels)
-            else:
-                for ch, ok, err in dispatch_alert(
-                        s, "critical", f"🔴 GPU OOM in {svc}", (detail or "")[:1500]):
-                    if not ok: print(f"notifier {ch} error:", err, flush=True)
-    except Exception as e:
-        print("notify_scan oom error:", e, flush=True)
-
-    # ── Uptime checks: per-check smart alerting (down / recovery / slow) ──────
-    try:
-        notify_uptime(s)
-    except Exception as e:
-        print("notify_scan uptime error:", e, flush=True)
+# Phase 3.3: moved to backend/notify/ — re-exported for backward compat
+from backend.notify import notify_scan
 
 def _uptime_down_reason(st):
     """Short, credential-safe reason for a DOWN check, taken from its last result."""
